Move mock settings into dedicated file
