+++ src/game/player.spec.ts
@@ -6,10 +6,11 @@ import { Vector } from "../utility/vector";
 import { Player } from "./player";
 import { KeyCode } from "./key-codes";
 import { GameContext } from "./game-context";
-import { GameSettings, getMockSettings } from "./game-settings";
+import { GameSettings } from "./game-settings";
 import { FastArray } from "../utility/fast-array";
 import { Projectile } from "./projectile";
 import { Obstacle } from "./obstacle";
+import { getMockSettings } from "../test/game-settings";
 
 const playerAnimationMock = {
     idle: [new AnimationFrame(1, 2, 3, 4)],
